requirements.txt
numpy==1.26.4
opencv_contrib_python==4.6.0.66
opencv_python==4.6.0.66
opencv_python_headless==4.9.0.80
paddleocr==2.7.0.3
pdf2image==1.17.0
Pillow==10.2.0
tensorflow==2.16.1
tensorflow_intel==2.16.1
tqdm==4.66.2
